Roughly enable record browsing
Allow browsing forward and backward on assumption that nameids
are consecutive.

Currently does not deal with the case where we go lower or
higher than the range of nameids.

Includes some commented-out code. This code currently causes
infinite loops. The intent was to be able to use separate
functions for iterating forwards or backwards, using the API
so that I don not have to assume consecutive nameids. I wasn't
a fan of this approach anyway, as it required an annoying
amount of prop passing -- though this can be mitigated a
little by moving the control panel up a level.