Better control of loading state
Spin on initial load

Thereafter, use context to lock the interface while records are
loading.

Not thrilled with using two flags for this, I'd rather use a
counter, but that didn't work out -- which does make me worry
about whether there might be timing issues here, but for an
interface it is probably good enough.