Initial support for "other" data and services for ratings
Add a new route to extras. This loads a tabbed page giving access
to data from other sources about both the person and their
services.

Need to figure out how I'm managing state
Currently a mix of prop drilling and useLocation, which is a
react-router feature. Need to understand how useLocation is
working, and whether I'm happy with how much state that is
exposing to absolutely everything.

Using a cobbled-together means to get the data from the database
for now.