Create datatable "superclass" to give me a reusable data table
I have a few tables of data to show which are quite similar.

The special case is the existing one of the pair of tables of
service records.

So create a generic datatable, then override it to provide the
servicetable.

The most complicated bit is providing the buttons: now we have
an extraRowControls parameter which takes a function which expects
to receive the current row params and an sx as its own params.

If extraRowControls is provided then the new datatable uses it to
generate extra buttons beyond the standard add above, add below
and delete options.