Update datagrid state in a new way
Working through this function seems to be much more reliable than
using onCellEditStop.

For example, using onCellEditStop does not give access to the new
value when editing is terminated by the user clicking out.

Note that we still have to return the new row to keep the internal
state of the grid consistent with the React state that is being
controlled.

I would much prefer the grid to behave reliably like a controlled
component, but so far this solution seems adequate.