Rearrange ratings index to use nameid, not officialnumber
Official number is a bit inconsistent. Occasionally can have
prefix that differs by case only, or can contain a comma and an
extra number, or can be NULL.

nameid is the catalogue ref and also (IIRC) the primary key, so
both makes sense in the interface and should be much more
consistent.

This does mean that missing offical numbers will not show up if it
is ever the case that the mapping from official number to item
number ever breaks down (e.g. if the item numbers are made
consecutive in the catalogue regardless of the official number in
the underlying record).

So this interface is suitable for finding cases where the
catalogue is 'aware' of missing official numbers, but might not be
suitable for finding missing official numbers in general.