Remove requirement that DataTable has a row
And also remove requirement that the row is one higher than the
array index.

This may always have been wrong in cases where the table was
made sortable.

Now DataTable has an identified primary key, which must not be
editable.

Id the prop positionalPrimary is also set then ordering will be
forced to match the primary key.

We no longer require that sort order match primary key order in
the array. But for the difference check in ServiceReconciler we
do still need that to be the case, so the row integrity check is
moved to there.

ServiceReconciler's additional buttons also still are written to
require that enforced relationship between row and array position.
I could likely rewrite it not to require that relationship, but as
we still have the integrity check here, there is no need to do so.