Fix differencemap when keys are not the same
The two services tables do not always have the same keys -- if I
add an empty row, it only has the row key, for example.

The inner loop is based on keys found in the outer loop -- which
means if a key exists in table2 but not in table1, we will not
detect this.

I could arrange things so that all keys are always present, but
seems more robust to have the differencer highlight keys that only
exist on one side.

Instead, just do the difference computation twice, swapping the
"inner" and "outer" roles of the tables, then merge the resulting
difference maps.

There is probably a more efficient way to do this, but this one
seems simple and robust.