Block navigation when there is dirty data
Navigation performed through react-router can now be blocked.
Only implemented for person.js, which will pop up a dialog if the
user tries to navigate away while any data is dirty.

Will *not* block navigation when moving between tabs.

As we are now tracking "dirtiness", we also use this for enabling
the various Enter buttons.

Adds zustand to the package for managing global synchronous state.
We use this to keep track of whether or not there is dirty data.