Get ratings structure from API
This makes much more sense than duplicating it in the front end.

I do manipulate the structure a bit in queries to make it as the
front end expects, but this is isolated in one place, at the point
where the data is acquired, so is fine.

Officer data still provided by the front end for now.