Typing and error checking for persontabledata
String vs numeric nature of data should now be respected.

Blank numeric fields are treated as being zero, and explicitly set
to 0 as soon as possible. They cannot be set to zero while the
field is being edited, because then it becomes impossible to delete
the contents of a field before entering a new value.

Annoying hack in dirty.js prevents the scope of these changes from
being limited to persontable. This is needed so that temporarily
blank fields do not in themselves enable the commit button.

There is probably a better way to do validation, but this way works.
It also permits entry of invalid values (e.g. impossible dates),
merely flagging them for the user.