useQuery corrections
Remove refetchOn*. This prevents stale data for being refetched in
each of these conditions. Because staleTime is Infinity, we are in
control of when data becomes stale, and when it becomes stale, we
always want it refetched.

Call invalidateQueries with an object with the relevant key as its
'queryKey' member, which appears to be the correct way to use this.
I think that previously this was invalidating all queries, not the
specific one that I was trying to invalidate.