Change piece bucket implementation to use a subscription
This dodges problems I was having the managing the asynchronous
state. I would rather understand the problems, but the time has
come to draw a line.

It does also make this temporary feature lift out more easily, so
it is a better way to do it for at least that reason.