Take snapshot before and after mutations and on leaving record
Actually at point of send (when an enter button is
pressed) and when new data is received within person
(detected by monitory query state and mutation state
for change) and at point of departure from the record
(provided this is done using buttons within the interface).

Show a screenshot before send if the relevant option is
selected in the menu bar.

If we cannot take a screenshot, still download the raw data.