Take snapshot before and after mutations
Actually at point of send (when an enter button is
pressed) and when new data is received within person
(detected by monitory query state and mutation state
for change).

Show a screenshot before send if the relevant option is
selected in the menu bar.

If we cannot take a screenshot, still download the raw data.