Return to running all existing service tables through translateToAPI
I had reduced this to just doing the first table, as we can only
write if the tables are identical.

Unfortunately, we use this same function to write any part of the
data, so if we write anything that is not a service table, we
also delete the right-hand table.

Reverts the part of 69f6c16e3 that had implemented this bad
decision. Instead, serviceRecordsMutate drops the right-hand
table before the translateToAPI call. This should be safe as a
serviceRecordsMutate call means that the user is actually writing
the service tables.