Adapt to tighter "not in bucket" response
The response now contains a body stating how many records were
returned -- primarily so that we have a response unique to this
application and know that our 404 was not caused by anything else.
Tighten it up as much as I can anyway, because I prefer an error
to alert me to surprising cases.