WIP
Reverted commit f9872d9cac3b0de473bb30709db964da9d589bfa requires
updates to the tests, as the test data then needs to contain
every field that real data would contain.

This is easy to manage for the pre-cooked .json files.

But some test data is generated. This is more complex because,
until now, I've been able to assume that all service fields are
also reflected in the UI. This is not the case for the officer or
source_id fields. I've begun to make changes for this in api.jsx,
but the changes may or may not be incorrect: certainly, more
work is needed.