Common scope for all mutations
We only call useMutation in one place. Here we set the same scope
for all mutations. This forces mutations to run in series.

This means that, if all data updates are completed when the
mutation ends, the next mutation will be gathering the updated
data -- in other words, mutations become atomic.