Use onSuccess callback instead of relying upon promise
Should work either way, but the docs really want you to use the
callbacks, which creates a bit more confidence that this behaves
as I expect. It probably is a bit easier to understand like this,
too.

Note that there are caveats around mutation callbacks in the docs,
but these apply to extra callbacks passed to the mutate function,
not to callbacks defined with useMutation.