Introduce LockedContext and implement for person data
Also tightens up things that should be disabled on load.

When a button is pushed, lock everything that can change the app
state until we return to a settled condition.

In principle not needed as we should anyway be locked out when in
loading states. But actually if button action implementations get
complex then that may cease to be true. And anyway much better to
be robust with this stuff.