Remove use of setTimeout
While I would very much like to be able to disable user input while
calculating the next state, this seems not to be how React wants to
work.

For the sake of having something reasonably likely to be correct
for the user, opt to trust the React paradigm (to the extent that
  I understand it) and allow user inputs to batch up as changes.

To be fair, in practice this is likely to rarely if ever be an
issue. The important thing is that syncrhonization with the server
blocks user input, and this is still the case.