Correctly update reconciled state when xcheck becomes disabled
Was not doing this via setState. Direct setState in a parent is
impossible during render but apparently can be achieved via
useEffect.

Potentially could have led to records being committed in the past
with an apparent state of Xcheck unset but an actual state of
Xcheck set -- but only if the user sets xcheck and then creates a
condition in which it is disabled, and then actually intends to
commit in a not-set state.