Do not refresh sync state after mutate
The sync state until now was discarded and replaced with the
server state after mutation.

This was a way of ensuring that the two were in sync and might
even be RESTful.

However, it is also a potential source of subtle bugs, as pending
changes to the sync state perhaps can be overwritten with the
update from the server.

The async state is still updated, so we can still see that the sync
state matches the async state, as expected (or, indeed, if this is
not the case).

Was relying on some transformations that happened at async-write
time to apply changes to the sync state on refresh. These should
now all be applied to the sync state at appropriate moments (two
such cases discovered, both within servicereconciler).